fncs from ckptserializer.cpp -> dmtcp_restart.cpp
* dmtcp_restart no longer depends on ckptserializer.o; nor any _real_XXX() fncs
